CLI: Move external entry points into new cli sub-package
