+++ ifsbench/cli/__init__.py
@@ -1,4 +1,4 @@
-# (C) Copyright 2018- ECMWF.
+# (C) Copyright 2020- ECMWF.
 # This software is licensed under the terms of the Apache Licence Version 2.0
 # which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
 # In applying this licence, ECMWF does not waive the privileges and immunities
@@ -6,5 +6,9 @@
 # nor does it submit to any jurisdiction.
 
 """
-Standalone scripts that provide the top-level entry-point to ifsbench tools.
+Command Line Interfaces (CLI) and utilities for ifsbench functionalities.
 """
+
+from ifsbench.cli.cli import * # noqa
+from ifsbench.cli.pack import * # noqa
+from ifsbench.cli.nml_diff import * # noqa
